@cbilz 2025-02-28
Would you mind sharing your benchmarking code? I'm interested in it because I've done some related experiments previously.