@rohlem 2025-03-01
> \[...\] by leveraging "add with carry" instructions. It is quite hard to get the compiler to properly generate them \[...\]

Status-quo has [a builtin `@addWithOverflow`](https://ziglang.org/documentation/master/#addWithOverflow), I think exactly for this purpose,
which is also used in the previous implementation (renamed `lladdcarrysmall` after your commit).
(3 summands currently require two `@addWithOverflow` in sequence.)
Have you tried an implementation using that builtin? Did it not generate the right instruction, or was the overall code quality worse?