@samy-00007 2025-03-01
> [...] I think exactly for this purpose,

Not exactly, unfortunately, `addWithOverflow` does not take a carry from a previous operation.
Currently, I have not managed to produce add with carry instructions using it (maybe it is possible, but I couldn't figure it out).

Having `addWithOverflow` and `subWithOverflow` take a carry as a third parameter could probably help, tho.