@rohlem 2025-03-01
> > [...] I think exactly for this purpose,
> 
> Not exactly, unfortunately, `addWithOverflow` does not take a carry from a previous operation.

While I also think a variadic `@addWithOverflow` to allow 3 parameters would be handy, you can still write the algorithm in status-quo, like the previous version does.
You just need to use 2 in sequence:
```
const r0_without_carry, r0_first_carry = @addWithOverflow(a[i + 0], b[i + 0]);
r[i + 0], r0_second_carry = @addWithOverflow(r0_without_carry, additional_carry_bit);
const r0_effective_carry = r0_first_carry | r0_second_carry;
```