@samy-00007 2025-03-01
Yes, but this does not generate add with carry instructions.
Your code, which is similar to what was used before, produce for every add:
```asm
        mov     r10, qword ptr [rdi + 8*rax]
        add     r10, qword ptr [rdx + 8*rax]
        setb    byte ptr [rbp - 24]
        setb    byte ptr [rbp - 17]
        movzx   r11d, byte ptr [rbp - 1]
        and     r11d, 1
        add     r11, r10
        setb    byte ptr [rbp - 16]
        mov     qword ptr [r8 + 8*rax], r11
        movzx   r10d, byte ptr [rbp - 16]
        and     r10b, 1
        mov     byte ptr [rbp - 25], r10b
        or      r10b, byte ptr [rbp - 17]
        and     r10b, 1
        mov     byte ptr [rbp - 1], r10b
```

My code generates 
```asm
; get the carry
        movzx   eax, byte ptr [rbp - 24]
        bt      eax, 0

; save the carry
        setb    byte ptr [rbp - 24]
        mov     qword ptr [rbp - 32], rax
```
at the start and end of each loop (ignoring the loop logic) and then this
```asm
        mov     r11, qword ptr [rdx + 8*rax - 16]
        adc     r11, qword ptr [rdi + 8*rax - 16]
        mov     qword ptr [r8 + 8*rax - 16], r11
```
for every add ([godbolt link](https://godbolt.org/z/hEYGa95e8))